+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/BindingGraphGenerator.kt
@@ -137,38 +137,36 @@ internal class BindingGraphGenerator(
 
     // Add aliases ("@Binds")
     val bindsFunctionsToAdd = buildList {
-      addAll(node.bindsFunctions)
+      addAll(node.bindsCallables)
       // Exclude scoped Binds, those will be exposed via provider field accessor
-      addAll(node.allExtendedNodes.values.filter { it.isExtendable }.flatMap { it.bindsFunctions })
+      addAll(node.allExtendedNodes.values.filter { it.isExtendable }.flatMap { it.bindsCallables })
     }
-    bindsFunctionsToAdd.forEach { (bindingCallable, initialContextKey) ->
-      val annotations = bindingCallable.annotations
-      val parameters = bindingCallable.ir.parameters(metroContext)
+    bindsFunctionsToAdd.forEach { bindingCallable ->
+      val annotations = bindingCallable.function.annotations
+      val parameters = bindingCallable.function.ir.parameters(metroContext)
       // TODO what about T -> T but into multibinding
       val bindsImplType =
         if (annotations.isBinds) {
           parameters.extensionOrFirstParameter?.contextualTypeKey
             ?: error(
-              "Missing receiver parameter for @Binds function: ${bindingCallable.ir.dumpKotlinLike()} in class ${bindingCallable.ir.parentAsClass.classId}"
+              "Missing receiver parameter for @Binds function: ${bindingCallable.function.ir.dumpKotlinLike()} in class ${bindingCallable.function.ir.parentAsClass.classId}"
             )
         } else {
           null
         }
 
-      val contextKey =
+      val targetTypeKey =
         if (annotations.isIntoMultibinding) {
-          IrContextualTypeKey.create(
-            initialContextKey.typeKey.transformMultiboundQualifier(metroContext, annotations)
-          )
+          bindingCallable.target.transformMultiboundQualifier(metroContext, annotations)
         } else {
-          initialContextKey
+          bindingCallable.target
         }
 
       val binding =
         Binding.Alias(
-          contextKey.typeKey,
+          targetTypeKey,
           bindsImplType!!.typeKey,
-          bindingCallable.ir,
+          bindingCallable.function.ir,
           parameters,
           annotations,
         )
@@ -176,12 +174,12 @@ internal class BindingGraphGenerator(
       if (annotations.isIntoMultibinding) {
         graph
           .getOrCreateMultibinding(
-            pluginContext,
-            annotations,
-            contextKey,
-            bindingCallable.ir,
-            annotations.qualifier,
-            bindingStack,
+            pluginContext = pluginContext,
+            annotations = annotations,
+            contextKey = IrContextualTypeKey.create(targetTypeKey),
+            declaration = bindingCallable.function.ir,
+            originalQualifier = annotations.qualifier,
+            bindingStack = bindingStack,
           )
           .sourceBindings
           .add(binding.typeKey)
@@ -207,6 +205,15 @@ internal class BindingGraphGenerator(
           bindingStack,
         )
       }
+    }
+
+    node.bindingContainers.forEach {
+      val typeKey = IrTypeKey(it)
+      graph.addBinding(
+        typeKey,
+        Binding.BoundInstance(typeKey, it.name.asString(), it),
+        bindingStack,
+      )
     }
 
     // Traverse all parent graph supertypes to create binding aliases as needed